refactor(artifacts): split file
